code cleaning and refactoring
